requirements.txt
Django>=5.2,<5.3
gunicorn>=21.2.0
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0 